Add a COMMIT_BODY option
Some GitHub actions can check for specific content in a commit message's body. For example a versionbot that checks for content like `Change-type: patch`.

This change allows the calling workflow to specify a string to use as the commit body. This string will be appended to the commit message, separated by two new lines.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>